+++ src/games/space-invaders/objects/bullet.ts
@@ -1,0 +1,42 @@
+/**
+ * @author       Digitsensitive <digit.sensitivee@gmail.com>
+ * @copyright    2018 Digitsensitive
+ * @description  Space Invaders: Bullet
+ * @license      Digitsensitive
+ */
+
+export class Bullet extends Phaser.GameObjects.Image {
+  private currentScene: Phaser.Scene;
+  private bulletSpeed: number;
+
+  constructor(params) {
+    super(params.scene, params.x, params.y, params.key);
+
+    this.initVariables(params);
+    this.initImage();
+    this.initPhysics();
+
+    this.currentScene.add.existing(this);
+  }
+
+  private initVariables(params): void {
+    this.currentScene = params.scene;
+    this.bulletSpeed = params.bulletProperties.speed;
+  }
+
+  private initImage(): void {
+    this.setOrigin(0.5, 0.5);
+  }
+
+  private initPhysics(): void {
+    this.currentScene.physics.world.enable(this);
+    this.body.setVelocityY(this.bulletSpeed);
+    this.body.setSize(1, 8);
+  }
+
+  update(): void {
+    if (this.y < 0 || this.y > this.currentScene.sys.canvas.height) {
+      this.destroy();
+    }
+  }
+}